fixup! Switch all signals/operations to batch-first

+++ nengo_dl/tests/test_learning_rules.py
@@ -6,6 +6,7 @@ import nengo
 from nengo.builder.learning_rules import SimVoja, SimOja, SimBCM
 import numpy as np
 import pytest
+import tensorflow as tf
 
 from nengo_dl import configure_settings, graph_optimizer
 from nengo_dl.learning_rule_builders import SimPES
@@ -26,7 +27,10 @@ def test_merged_learning(Simulator, rule, weights, seed):
     # make sure that works OK
     dimensions = 2
     with nengo.Network(seed=seed) as net:
-        configure_settings(planner=partial(graph_optimizer.tree_planner, max_depth=10))
+        configure_settings(
+            planner=partial(graph_optimizer.tree_planner, max_depth=10),
+            dtype=tf.float64,
+        )
 
         a = nengo.Ensemble(3, dimensions, label="a")
         b = nengo.Ensemble(3, dimensions, label="b")
